Split testerina test cases to avoid build failure
