+++ tests/testerina-integration-test/src/test/java/org/ballerinalang/testerina/test/AnnotationsTest.java
@@ -15,7 +15,6 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-
 package org.ballerinalang.testerina.test;
 
 import org.ballerinalang.test.context.BMainInstance;
@@ -24,10 +23,10 @@ import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
 
 /**
- * Test class to test predeclared langlibs using a ballerina project.
- * @since 2.0.0
+ * Test class to test positive scenarios of testerina using a ballerina project.
  */
-public class ImportTest extends BaseTestCase {
+public class AnnotationsTest extends BaseTestCase {
+
     private BMainInstance balClient;
     private String projectPath;
 
@@ -37,9 +36,8 @@ public class ImportTest extends BaseTestCase {
         projectPath = projectBasedTestsPath.toString();
     }
 
-    @Test()
-    public void testBeforeAfter() throws BallerinaTestException {
-        balClient.runMain("test", new String[]{"predeclared-import-tests"}, null,
-                new String[]{}, projectPath);
+    @Test
+    public void testAnnotations() throws BallerinaTestException {
+        balClient.runMain("test", new String[]{"annotations"}, null, new String[]{}, projectPath);
     }
 }
