+++ tests/testerina-integration-test/src/test/java/org/ballerinalang/testerina/test/AssertionErrorsTest.java
@@ -15,7 +15,6 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-
 package org.ballerinalang.testerina.test;
 
 import org.ballerinalang.test.context.BMainInstance;
@@ -24,10 +23,10 @@ import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
 
 /**
- * Test class to test predeclared langlibs using a ballerina project.
- * @since 2.0.0
+ * Test class to test positive scenarios of testerina using a ballerina project.
  */
-public class ImportTest extends BaseTestCase {
+public class AssertionErrorsTest extends BaseTestCase {
+
     private BMainInstance balClient;
     private String projectPath;
 
@@ -37,9 +36,15 @@ public class ImportTest extends BaseTestCase {
         projectPath = projectBasedTestsPath.toString();
     }
 
-    @Test()
-    public void testBeforeAfter() throws BallerinaTestException {
-        balClient.runMain("test", new String[]{"predeclared-import-tests"}, null,
-                new String[]{}, projectPath);
+    @Test
+    public void testAssertDiffError() throws BallerinaTestException {
+        balClient.runMain("test", new String[]{"assertions-diff-error"}, null, new String[]{},
+                projectPath);
+    }
+
+    @Test
+    public void testAssertionErrorMessage() throws BallerinaTestException {
+        balClient.runMain("test", new String[]{"assertions-error-messages"}, null, new String[]{},
+                projectPath);
     }
 }
