--- tests/testerina-integration-test/src/test/java/org/ballerinalang/testerina/test/BasicCasesTest.java
@@ -1,112 +1,0 @@
-/*
- * Copyright (c) 2020, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- *
- * WSO2 Inc. licenses this file to you under the Apache License,
- * Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-package org.ballerinalang.testerina.test;
-
-import org.ballerinalang.test.context.BMainInstance;
-import org.ballerinalang.test.context.BallerinaTestException;
-import org.testng.annotations.BeforeClass;
-import org.testng.annotations.Test;
-
-/**
- * Test class to test positive scenarios of testerina using a ballerina project.
- */
-public class BasicCasesTest extends BaseTestCase {
-
-    private BMainInstance balClient;
-    private String projectPath;
-
-    @BeforeClass()
-    public void setup() throws BallerinaTestException {
-        balClient = new BMainInstance(balServer);
-        projectPath = projectBasedTestsPath.toString();
-    }
-
-    @Test
-    public void testAssertions() throws BallerinaTestException {
-        balClient.runMain("test", new String[]{"assertions"}, null, new String[]{}, projectPath);
-    }
-
-    @Test
-    public void testAssertDiffError() throws BallerinaTestException {
-        balClient.runMain("test", new String[]{"assertions-diff-error"}, null, new String[]{},
-                projectPath);
-    }
-
-    @Test
-    public void testAssertionErrorMessage() throws BallerinaTestException {
-        balClient.runMain("test", new String[]{"assertions-error-messages"}, null, new String[]{},
-                projectPath);
-    }
-
-    @Test
-    public void testAssertBehavioralTypes() throws BallerinaTestException {
-        balClient.runMain("test", new String[]{"assertions-behavioral-types"}, null, new String[]{},
-                projectPath);
-    }
-
-    @Test
-    public void testAssertStructuralTypes() throws BallerinaTestException {
-        balClient.runMain("test", new String[]{"assertions-structural-types"}, null, new String[]{},
-                projectPath);
-    }
-
-    @Test
-    public void testAssertSequenceTypes() throws BallerinaTestException {
-        balClient.runMain("test", new String[]{"assertions-sequence-types"}, null, new String[]{},
-                projectPath);
-    }
-
-    @Test
-    public void testAnnotationAccess() throws BallerinaTestException {
-        balClient.runMain("test", new String[]{"annotation-access"}, null, new String[]{},
-                projectPath);
-    }
-
-    @Test
-    public void testJavaInterops() throws BallerinaTestException {
-        balClient.runMain("test", new String[]{"interops"}, null, new String[]{}, projectPath);
-    }
-
-    @Test
-    public void testBeforeAfter() throws BallerinaTestException {
-        balClient.runMain("test", new String[]{"before-after"}, null, new String[]{}, projectPath);
-    }
-
-    @Test
-    public void testBeforeEachAfterEach() throws BallerinaTestException {
-        balClient.runMain("test", new String[]{"before-each-after-each"}, null,
-                new String[]{}, projectPath);
-    }
-
-    @Test(dependsOnMethods = "testBeforeAfter")
-    public void testDependsOn() throws BallerinaTestException {
-        balClient.runMain("test", new String[]{"depends-on"}, null, new String[]{}, projectPath);
-    }
-
-    @Test(dependsOnMethods = "testDependsOn")
-    public void testAnnotations() throws BallerinaTestException {
-        balClient.runMain("test", new String[]{"annotations"}, null, new String[]{}, projectPath);
-    }
-
-    @Test
-    public void testIsolatedFunctions() throws BallerinaTestException {
-        balClient.runMain("test", new String[]{"isolated-functions"}, null, new String[]{},
-                projectPath);
-    }
-
-}
